requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
blis==0.7.7
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-log==0.1.8
cryptography==35.0.0
cymem==2.0.6
distro==1.7.0
elasticsearch==7.13.4
fastapi==0.68.2
Flask==1.1.4
Flask-Cors==3.0.10
frozenlist==1.3.0
future==0.18.2
h11==0.13.0
idna==3.3
immutables==0.17
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.0
MarkupSafe==2.0.1
marshmallow==3.7.1
mindmeld==4.5.0
multidict==6.0.2
murmurhash==1.0.7
mypy==0.950
mypy-extensions==0.4.3
nltk==3.7
numpy==1.21.6
passlib==1.7.4
plac==1.1.3
preshed==3.0.6
py==1.11.0
pycountry==22.3.5
pycparser==2.21
pydantic==1.9.1
PyJWT==2.3.0
pymongo==4.0.2
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.26.0
requests-toolbelt==0.9.1
scikit-learn==0.19.2
scipy==1.7.3
six==1.16.0
sklearn-crfsuite==0.3.6
spacy==2.3.7
srsly==1.0.5
starlette==0.14.2
tabulate==0.8.9
thinc==7.4.5
tomli==2.0.1
tqdm==4.64.0
typed-ast==1.5.3
typer==0.4.1
typing-extensions==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
wasabi==0.9.1
webex-skills==2.0.15
webexteamssdk==1.6
Werkzeug==1.0.1
yarl==1.7.2
zipp==3.8.0